third_party/blink/web_tests/virtual/single-import-map/external/wpt/import-maps/acquiring/dynamic-import-expected.txt

This is a testharness.js-based test.
[FAIL] After a dynamic import(), import maps work fine
  assert_array_equals: expected property 0 to be "log:B" but got "log:A" (expected array ["log:B"] got ["log:A"])
Harness: the test ran to completion.
